248_Deformable_Transformers_for_VIS/requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.12
cycler==0.11.0
fonttools==4.30.0
idna==3.3
jsonpatch==1.32
jsonpointer==2.2
kiwisolver==1.4.0
matplotlib==3.5.1
numpy==1.22.3
opencv-python==4.5.5.64
packaging==21.3
Pillow==9.0.1
pycocotools==2.0.4
pyparsing==3.0.7
python-dateutil==2.8.2
PyYAML==6.0
pyzmq==22.3.0
requests==2.27.1
scipy==1.8.0
six==1.16.0
torchfile==0.1.0
tornado==6.1
tqdm==4.63.1
typing-extensions==4.1.1
urllib3==1.26.8
visdom==0.1.8.9
websocket-client==1.3.1
yacs==0.1.8
imgaug==0.4.0
timm==0.5.4